expressFormat use originalUrl if available
When using nested routers in e.g. express, the router modifies req.url to only contain the router relevant url part. For that matter, when logging req.url, you wouldn't see the complete original url. Express adds req.originalUrl which contains the url that was called by the client. If this property is available, it should be used over req.url in the log format